Add basic API call for Group to check if a group exists
Add a method to the AnVILAPISession to check if a group exists on
AnVIL. Add a method to the Group model to make the API call using
the session. Add a single test to check a successful response code.
More tests will need to be added to handle invalid response codes.