Handle error response codes for the Group anvil_exists method
Modify the AnVILAPISession class so that the `get` method is not
overridden and instead has a `_get` method that calls the superclass
`get` method, checks status codes, and then raises exceptions where
appropriate. This is to make the tests work -- previously, when
mocking the `AnVILAPISession.get` method, none of the exceptions
were raised because that method was called. There ies probably a
better way to handle this, and I will look into it, but at least the
tests work now.