Restructure AnVIL API code
Separate the code for working with the AnVIL API into two classes:
a Session class that subclasses AuthorizedSession (as before) and
handles the get/post/etc. requests and checking responses for error
codes. Then add an API Client class that has a class variable with
the authenticated session. That way, the same session will be shared
by all instances of the client.