Add an integration test for the AnVIL API code
This test actually works with AnVIL directly instead of using
mocking to make the API calls. Update pytest.ini so that this
test is not run by default. It can still be run with
`pytest anvil_project_manager/tests/integration/test_anvil_api.py`

There is probably a better way to do this and still have project-wide
test discovery.