Add API and model methods to check if a billing project exists on AnVIL
Add a method to the API to return information about a billing group
and corresponding integration tests. Add a method to the BillingProject
model to check if a billing project exists on AnVIL. Right now I
don't expect to be creating a billing project on AnVIL using this
interface, so there is no way to create one, but that could change.