Call the AnVIL API when creating or deleting a group
When creating a group on the django site, first make an API call
to create it on AnVIL. If that succeeds, then save the object.
Similarly, when deleting a group, first try to delete the group via
the AnVIL API. If that succeeds, then delete the object.

In some cases, the API does not appear to be returning the documented
code. For those cases, I have added tests that skipped until we
figure out how we want to handle them.