Remove ability to update group membership roles
In AnVIL you can be both a MEMBER and an ADMIN of a group, so to
change a user from a member to an admin, you need to delete them
as a member and then add them as an admin. (Otherwise they will
also remain a member.) To do this on AnVIL we would need to make
two API calls, one to remove them as a member and one to add them
as an admin. How would we handle a case where one call fails but
not the other?

 If we decide we need this functionality, we can figure out how to
implement the API calls and account for failures.