Mock the authorized session credentials object in tests
Previously, tests were failing when you couldn't authenticate with
google.auth.default because the tests were actually running the
bit of code to try to authetnicate. Fix this by mocking the
credentials object that is passed to AuthorizedSession. That way,
the tests assume we have authenticated and proceed happily along.