Add a class method to import a Workspace from AnVIL
This method checks if the workspace exists on AnVIL, if the app is
an OWNER of it, and then creates a new instance and saves it. If
necessary, it also creates a billing project if the workspace's
billing project is not already in the Django database. Because it
needs to create two objects, transactions are used.