Add a view to import a workspace from AnVIL
As part of this, add a form class and tests for it. This can't be
a model form because in some cases, we need to create the record for
the associated billing group, and in other cases it will exist. Add
tests for this form. Add tests, a template, and a url for the view.
Add a link to this view in the navbar.