Do not allow groups used as auth domains to be deleted
Add an explicit through table (sigh! I had just removed this) because
it allows us to specify the on_delete behavior. We do not want to
be able to delete a group that is used as an auth domain on AnVIL.
Add tests of the models to ensure that the expected behaivior is
happening (error when trying to delete a group used as an auth
domain). Add logic to the GroupDelete view to redirect with a message
when attempting to delete a group used as an auth domain.