Check is_managed_by_app status when creating group memberships
Do not allow a GroupGroupMembership to be created if the
is_managed_by_app flag of the parent group is False, and do not
allow a GroupAccountMembership to be created if the flag of the
group is false. This means that we don't have admin privileges for
that group on AnVIL, so we can't add members to it. Add a form with
custom validation for the parent_group field, and modify the
GroupGroupMembershipCreate and GroupAccountMembershipCreate views
to use these new form. Add tests for both the forms and the modified
views' behavior.