Add validators for case-insensitive Account, Groups, and BillingAccount
Add custom validators to check if the account email already exists
(case-insensitive) or if a group or billing project name already
exists (case-insensitive). This means that if a group named "foo"
exists, the clean method of a group named "FoO" will raise a
validation error with a nice display on the form. It does not prevent
the model from being saved to the database, though. Add a migration
for these changes.