Verify workspace name case insensitivity in clean method
For workspaces, we couldn't use a custom validator on the name field
because you can have workspaces with the same name in different
billing projects. Instead, check for case insensitivity in the
model clean method, because we need to check two fields together.
As with the other validators, this won't prevent saving to the
database but it does give a nicer form error.