Add a BillingProject field to indicate if the app is a user
The app can either be users of a billing project (which means that
we can create workspaces in that billing project), or it can import
a billing project that it isn't a user of when importing a workspace.
We will need to track if we are users in a billing project to
safeguard against trying to create workspaces in billing projects
that we aren't part of.