Modify Workspace.anvil_import to use BillingProject.anvil_import
This allows us to set the billing project has_app_as_user field
for billing projects associated with workspaces. In some cases we
may be users of the billing project, and in others we will not be
users. Add tests for the workspace anvil_import method and associated
views that use it to confirm that has_app_as_user is set
appropriately. Also modify tests where necessary to add mocked
responses to the billing project API url.