Limit billing project choices to those we have access to
When creating a workspace, do not allow workspaces marked as
has_app_as_user=False because we are not users in that billing
project, and therefore can't create workspaces in it. Add a new form
to limit the billing project choices, and add tests in the views to
check that you can't create a workspace in an invalid billing
project.