Set on_delete for GroupGroupMembership
Do not allow groups to be deleted if they are part of another group
because it is currently not allowed by AnVIL. This is odd behavior,
so I have opened a support ticket but am not sure what will come of
it. Note that parent groups can be deleted, so that is allowed by the
app. Add tests for this behavior, and modify the ManagedGroupDelete
view to redirect with a message when attempting to delete a child
group.