Remove user from groups when deleting an account from the app
Remove a user from all groups when deleting the account from the app.
This has to make multiple API calls, and if any fail, it will report
an error to the user and not delete any of the relationships from
the app. Unfortunately someone will need to go and fix the issue
manually, at least for now.