Verify that group was actually deleted on AnVIL
The AnVIL API is buggy for deleting groups and often returns a
successful response code when the group could not actually be
deleted. Some of the potential cases could occur in this app (e.g.,
if someone adds one of our groups to their group as a member).
Therefore, the anvil_delete method for ManagedGroups and the
ManagedGroupDelete view need to handle this possibility. Hopefully
this will not actually happen, but it's better to handle it in case
it does.

To implement this, make a second API call in the anvil_delete method
to check if the group still exists on AnVIL. If it does, raise a
custom exception indicating that the group could not be deleted.
In the ManagedGroupDelete view, handle this exception by displaying
a message to the user and redirecting to the group detail page.

I have opened a couple support tickets on AnVIL about this.
https://support.terra.bio/hc/en-us/requests/274901
https://support.terra.bio/hc/en-us/requests/279828