Require login for all views in app other than the index
Add LoginRequiredMixin for all views. Update tests so they create a
test user and log it in (when using the client) or attach it to the
request (when using the RequestFactory).  Add a test for all views
ensuring that a user is redirected if they are not logged in. For
views that process both get and post requests, add two tests.