Show message when a billing project is successfully imported
Add a success message when a billing project is successfully imported.
Note: to test this functionality, we have to add the Session and
Messages middleware to RequestFactory requests. The suggestion in
Two Scoops of Django causes a deprecation warning since Middleware
was updated in Django 4.0. Is it better to do this or to just use
the Django testing client?