Add skipped tests for foreign key relationship history
These tests highlight some issues with the history and foreign keys
when the related object is deleted. The objects appear in the
history, but can't be loaded because the related object does not
exist anymore. Add some tests that will fail because of this, and
then skip them. I'm not sure if this will end up being a problem.