Show different components on template based on Account status
Modify the context data to indicate if an account is inactive and
if the reactivate or deactivate buttons should be shown. Modify the
template to use these flags to 1) selectively show an alert if an
account is inactive and 2) show the proper deactivate/reactivate
buttons based on status (only active accounts should show the
deactivate button and only inactive accounts should show the
reactivate button).