Modify tests failing in MariaDB due to collation differences
It appears that MariaDB returns the queryset in a different order
(alphabetically), where SQLite returns them in the creation order.
When creating workspaces with two auth domains, the groups were
sometimes in a different order than what the responses package was
looking for. This order doesn't matter, but it caused the tests to
fail. Fix them by specify names when creating the groups and make
sure they are in alphabetical order.