Raise an exception when sharing a workspace with a non-existent group
Previously, if a group existed in the app but not on AnVIL, the
app would happily share a workspace with that group. Modify the
anvil_create_or_update method to check the "usersNotFound" field
of the response body, which is populated when a group is not found.
Note that this is unlikely to happen in practice but better to fix
it here anyway.