Use the name as a slug in the url for BillingProject
Instead of using the pk, use a slug in the url for BillingProject.
This can just use the "name" field, which is already a SlugField
(although it is case-sensitive). To allow the import url to still
work, move the detail view url after the other urls. Otherwise,
when going to the import url, Django tries to call the detail view
for a billing project named "import". It might be better to prefix
the detail view with "detail" so there no chance of collisions.