Add an abstract base model and default model for workspace data
Add an abstract base model that projects extending this app can
inherit to create their custom WorkspaceData model. As part of this,
to make the app work without having to customize it, add a
DefaultWorkspaceData model in the main app. Unfortunately this
creates an extra table (for DefaultWorkspaceData) and field
(workspace.defaultworkspacedata) if a project is implementing a
custom workspace data model. Not sure if this is the best way to do
this, or if we should just require projects implementing the app to
create their own WorkspaceData model and not provide a default.

Update docs to reflect these changes.