Add workspace as a field to the workspace_data_form
Modify the WorkspaceCreate and WorkspaceImport views to use set
workspace when getting the workspace_data_form. This means that
the WorkspaceImport view needs to make an API calls before running
the workspace_data_form.is_valid() method, which is unfortunate but
I'm not sure if there's a way around it. Modify docs to state that
the workspace field is required.