Register adapters specified in settings in appconfig.ready()
Add a setting where the user can specify the list of workspace
adapters that they want to use in this project. Then register the
adapters in the appconfig.ready() method so that the registry knows
about them. Trying to register them in an app's adapters.py file
doesn't work, because the adapters.py file is never imported anywhere.
We could get around this by autodiscovering files named 'adapters.py'
or 'workspace_adapters.py' (similar to how the admin registration
works), but having a setting is more flexible for users and more
explicit.