Update WorkspaceCreate to use the adapter registry
Pass the workspace type to create as a url parameter, then look
up the correct adapter in the registry using that paramter. Add a
mixin to handle this looking and set an instance variable for the
adapter. Modify the view to use the mixin and get the adapter from
the registry. Modify tests as necessary. As a side effect, I had to
remove the workspace dropdown from the navbar because the previous
"new" url was no longer valid without passing a workspace type. In
the long run we'll need to add some way of getting the registered
workspace adapters for use in the navbar and other templates.