Handle different workspace types in templates
Add a context processor that adds the set of registered workspace
adapters and their names to the context. The templates can use
that object to iterate over the workspace types and add links. To
do this, require that an adapter has a "name" field and add a method
to the registry to return a dictionary of registered adapter names.
Show the different workspace types in the dropdown navbar under
"workspaces", and have a different card for each different type of
workspace.