Modify navbar in templates
Use a non-fluid container so it works better with our sites. Add a
block for extra nav items in the ACM navbar. Move the login/logout
from the ACM navbar to a separate navar in the example site. This
way, other projects can use the navbar directly even if the login/
logout urls are different. Treat the ACM navbar as an "extra", second
navbar -- this is closer to how we'll be using the navbar in the
separate consortium Django projects.