Move version number to __init__.py
Store the version number of the package in __init__.py. This means
that the package has a version numbera ssociated with it in python
when evaluating "anvil_consortium_manager.__version__". It also
means that the setup.cfg and sphinx docs can get the version directly
from the same place, so version is only tracked in one place. No
more updating two files!