Add (failing) tests for AccountLink view
These are the main tests that we want to have for the AccountLink
view. Most of them are failing, which will require some changes to
the AccountLink view. Note that I had to change a couple other
things in the code: add an unrestricted view to redirect to in tests
and add some settings to the test settings file.