Change on_delete to PROTECT for Account.user foreign key
We still need to figure out what will happen when an account is
deleted, and how we will want to handle it. It is also unlikely
that an account will be deleted, and there will be some offboarding
procedure. For now, don't allow accounts to be deleted if they are
linked to a user. We can change it later in the future if we want.