Remove can_be_verified field from UserEmailEntry
It's better to check if an Account exists with the given email
or user instead of storing a flag in the database, since CC staff
could create an Account with the same email before anyone completes
the email verification process. If that happened, the email would
be flagged as "can_be_verified" but creating the account would fail.