Various bug fixes
Always compare lowercase emails. Remove the service account as a
member when auditing group membership -- sometime it is also listed
as a member. Do not require that the service account is a direct
owner of a workspace -- in some cases, it is part of a group that is
the owner.