Fix audit bug where app is both a member and admin of a group
When the app was both a member and an admin of a group, the Managed
Group audit reported a couple errors: a not_in_app problem and an
error about the app having a different role. Fix the anvil_audit_membership
method to not return these errors in this case.