Register the TestWorkspaceAdapter in custom template test
The test_render_custom_template_name test was failing because the
TestWorkspaceAdapter was not registered with the app. Use the same
strategy as for the WorkspaceCreate tests where we unregister the
default adatper and register the test adapter, and add a tearDown
method to clean up and register the default again.

Note: I'm not sure why this test worked in the CI but not on my machine?