Add view to share workspace with a group using workspace in url
Add a view to share a workspace with a group where the workspace
itself is specified in the url. This is similar to the previous
commit, which creats a new WorkspaceGroupSharing object by pulling
both the workspace and the group from the url. In this case, only
the workspace is pulled from the url.