Fix buggy duplicate Workspace test
This test was checking that a specific validation error was raised
in a specific circumstance, but a second validation error was being
raised for other reasons. Modify the test to check that 1) only the
one exception was raised and 2) that it was the correct exception.
Then fix the previous part of the test to pass the new checks.