Add functionality to update a workspace and workspace data
Add a new form for updating a workspace. Add a view, template, and
new url for updating the workspace and associated workspace data. Add
tests. As part of this, move the clean method that checks for an
existing workspace with the same case-insensitive billing project
and name to the WorkspaceCreate form instead of the Workspace model.