Add an anvil_clone method for a Workspace
This method clones an existing workspace on AnVIL, but does not
create or save a record of the new workspace in the app. This is
because we need to also specify the workspace type, and that can't
easily be done via this method. Therefore, any methods calling
anvil_clone should also call somethign like anvil_import or otherwise
create the new workspace plus workspace data in the app.