Add ability to clone a workspace
Add a form for cloning a workspace and tets for that form. Add a
view, url, and template for cloning a given workspace into a
specific workspae type. Add a button to the workspace detail page
for cloning that workspace. Modify the anvil_clone method for
Workspace to not check if a workspace already exists; all this
method does now is make the API call to AnVIL with the proper auth
domains. This is because the view to clone the workspace saves the
object and then calls anvil_clone.