Add a check for protected groups in delete view
Previously, the app allowed a group to be deleted on AnVIL before
checking if it could be deleted in the app (e.g., due to a protected
foreign key in another model). Add a test to check that this can't
happen and fix the view by using the django "Collector".