Move entry point to the API client class
Based on conversations with Terra support, we should be calling
APIs other than the api.firecloud.org api, which is the one that
appears in their documentation. As prep for this bigger change,
move the entry point definition to the API client class and rename
it to firecloud_entry_point.