Use SAM API for deleting groups
Note that this API appears to fix many of the issues that exist
with the group delete API - in particular, the API no longer returns
a correct response code when the group is not actually deleted (in
specific cases). Firecloud API used to return successful responses
whne the group could not be deleted.