Update anvil_exists method to use SAM API
Add a new method to the anvil_api client that calls the SAM API
to get a group's email. This is the closest analog to the get_group
method that calls the Firecloud API. Modify the anvil_exists method
for ManagedGroups to call the get_group_email method, and update
tests as necessary.