Don't run WorkspaceCreateForm clean check for Django 4.1 on mysql
Django < 4.1 did not check model constraints in the clean method,
and Django 4.1+ does. A couple things are going on here:
1. The Workspace model has a unique constraint on billing_project.name
   and name.
1. Django 3.2 does not have a case-insensitive unique constraint, so
   I added a custom check to the WorkspaceCreateForm clean method.
2. mysql/mariadb has a default case insensitive collation.

Now in Django 4.1+, the model constraints are checked during the
clean method. When running on mysql/mariadb (with case-insensitive
collation) and Django 4.1, the constraint check fails and a
ValidationError is raised from the model full_clean() method. In
sqlite and Django 4.1, the constraint check does not fail because
sqlite does not have case-insensitive collation (by default?). So,
add an if statement to skip this check in the clean method if both
the mysql/mariadb and the Django 4.1+ conditions are met. This relies
on the mysql/mariadb collation being set to case-insensitive by the
user, so it's maybe not the ideal solution.