Remove G kwarg from get_graph and add documentation
We no longer need to pass G to the get_graph method for a ManagedGroup
instance. Instead, this method can just create the graph itself.
I think this happened because I wanted to use the function recursively,
but I had to use the recursive add_parents and add_children methods
instead.