Verify lock status when auditing workspaces
Add a new allowed error to the WorkspaceAuditResults class for when
the lock status in the app doesn't match what's on AnVIL. Check the
audit status in the Workspace.anvil_audit method. Modify tests to
return the isLocked attribute for workspaces, and add new tests to
check that the error exists when lock status doesn't match.