Add a view mixin to add registered workspace adapters to context
This mixin takes care of adding registered workspace adapters to the
context data. Use this mixin in the WorkspaceDetail and the
WorkspaceLandingPage views, since they both need to access the set
of registered workspace adapters. Update templates and tests for
the new mixin, which is slightly different than how existing views
added info about registered workspaces to the context.