Handle group managed by app that does not exist in AnVIL in audit
Add a test is intended to capture the case where a group in the app
is marked as "is_managed_by_app" but the app is not a member of the
group on AnVIL (either as as a member or an admin). With the previous
changes, this would have returned a correct audit, but it was actually
incorrect.