Try setting up django-tables2 tables for emailed audit report
This is *broken*. Create two tables, an ErrorsTable and a NotInAppTable,
that will hold results from the run_anvil_audit report. Add a new
template that can be used to generate html that can be sent as part
of the email. Render that template with the audit results and the new
django-tables2 tables, and include that in the email that is sent by
the management command.

Unfortunately, this is not working -- django_tables2 requires a request
object to use the render_table template tag, but there is no request
in a management command.

One (commented out) alternative is to use json2html, but that makes
it harder to add links - it would be ideal to have links in the errors
table that point to the detail page for the object in question.