Add (relative?) link in ErrorsTable
Modify the to_json method to return the instance instead of the
string of the instance for the errors element of the report. Add
the linkify keyword to the instance field of the ErrorsTable. Note
that this adds a relative link, but I'm not sure how to make it add
the absolute link.