Include links in audit report tables
Modify the audit code such that links can be included in audit report
emails. This change introduces a new requirement that any project
using the app is also using the sites framework. Use the sites
framework in the run_anvil_audit management command to get the
current site, which is used to construct the url to the instance
without a request object. Modify tests as necessary to handle using
the sites framework instead of the testserver domain.