Include model name in email html title/h1
Include the model name in the h1 field of the email audit report
template. Obtain the model name from the model's meta object, but
use the object_name field instead of the verbose_name_plural. This
makes it nicer to put in the header. Modify tests checking the output
to check for eg "BillingProject... ok!" instead of "billing projects... ok!"