Fix tests that change site domain
There were tests that modified the domain on the default SITE_ID.
This caused one of the tests that used the default example.com site
to fail, since it assumed that the domain was example.com and somehow
it had been changed to foobar.com (by a previous test?). I find this
confusing since I thought it would have been handled in with
transactions, but apparently not? Fix this by overriding the SITE_ID
setting for these two tests and creating a new Site with a new domain
name.