Handle TestCase.assertQuerysetEqual deprecation warning
Django's TestCase class deprecated the assertQuerysetEqual method
in favor of assertQuerySetEqual. To handle testing on multiple versions
of Django, when running tests if the Django version is < 4.2, redefine
TestCase and add an assertQuerySetEqual method that calls the
assertQuerysetEqual method. This is similar to handling the change
in DeleteView in Django 4.0.