Fix not in app table in audit report emails
The NotInApp table in the audit report emails just showed dashes
for all records that aren't in the app. This happened because the
export results for "not_in_app" are a list of strings, not a dictionary
as expected by django-tables2. Modify this by adding a render_instance
method to the table, which just returns the string.

Closes #367