Move the "if q" statement to adapter autocomplete methods
The reason is that the adapter filtering should *always* be used,
whether or not there is a q parameter passed to the autocomplete.
With the previous code, this caused issues when a forwarded value
was passed but no query was entered. This caused the forwarded
value to be ignored and all records were returned.