Update WorkspaceDelete tests for WorkspaceDetail context change
The WorkspaceDelete view redirects to the detail view for a workspace
in some cases. The tests were not creating an associated workspace
data object, so in the tests that were checking redirection, the
WorkspaceDetail workspace_data_object context could not be populated.
Instead, create a DefaultWorkspaceData object in the tests, which
correctly creates both a workspace and an associated workspace data
object.