Refactor anvil audit code for ManagedGroupMembership
Change the auditing strategy such that auditing will no longer be
a method/classmethod of a model. Instead, it will be housed in a
separate source file and be its own classes.

This commit also has the base classes that will be used for audit
classes for other model class audit methods.