Add a custom filter form to display inputs in a single row
This custom form uses the crispy_form package's layout objects to
manipulate the placement and apperance of forms. In particular, it
displays the form fields on a single line, including the submit
button, with floating labels. This takes up a smaller footprint
and is less intrusive compared to the previous standard way of
displaying the form.