Add views tests for limited_view permission
Add tests for users with the limited_view permission. Right now,
all the tests check that the views raise 403 forbidden errors, since
the app is not updated to use these views. In the future the tests
can be changed to expect new behavior, when the app is expanded to
allow these users to see a subset of information.