Update the AccountList views to use the adapter filter
Add an AccountAdapterMixin, which sets the adapter attribute of
the view class. It also contains reusable methods to get information
from the adapter (e.g., the list class and the filterset class).
Use this mixin for all of the AccountList views, which uses the
list_filterset_class defined by whichever account adapter is set
by the project settings file.