Do not give audit errors for deactivated accounts not in a group
The ManagedGroupMembershipAudit currently gives an error if there
is a deactivated account that is in a group in the app, but is not
in the group on AnVIL. This is expected because the app stores the
record of what groups the account was in (so there is a record for
the GroupAccountMembership), but removes the account from all groups
on AnVIL (so the account email do not appear in the list of members/
admin returned by the AnVIL API). This should not raise an audit
error.