Modify ManagedGroupAudit to ignore non-admin groups not in app
The ManagedGroupAudit should not report "NotInApp" errors for groups
that are not in the app and where the service account is only a
member. This happens in the case where a group from the app has
been added to a non-app managed group that the app does not know
about (especially for GREGoR release workspaces).