Optionally allow ManagedGroupAutocomplete to filter on is_managed_by_app
The ManagedGroupAutocomplete method did not return groups that were
not managed by the app, which means that we cannot add groups not
managed by the app to other groups or share a workspace with a group
not managed by the app. This is not the intended behavior. Modify
the autocomplete method to have a forwarded value that specifies
whether to return only groups that are managed by the app.