Optionally store environment variables in a .env file
Since virtual environments are defined by hatch, we don't want to
set the environment variable for the service account file directly
in the hatch env definition, since the file is different for every
developer. Instead, modify the example project settings to read the
service account file from a .env file (optionally). This requires
adding the django-environ package to the dev dependencies.