Run tests on mysql and sqlite3
In pyproject.toml, set env variables in hatch environments to specify
whether we are using the MySQL or sqlite3 backend in tests. Note:
one test is breaking when running in MySQL; need to look into this.
Also, I expect CI to fail because I have not set up the MySQL/MariaDB
service in the github action workflow yet.