Specify overriden SITE_ID setting differently in tests
One test was failing in only mysql when run with pytest, but not
with unittest/manage.py test. This failure happened because the
pytest settings are to reuse the database between tests; in mysql,
that means that the autoincrement value for id columns is not reset
between tests. Therefore, hard coding SITE_ID=2 in the tests caused
the second test overriding SITE_ID to fail, since the new site
created in the second test had an id of 3. Change this to setting
the SITE_ID setting using a context manager to the value of the of
the site that was just created instead of using the override_settings
decorator. Now they pass!