Rename permissions in models and migrations
Add a migration to create the new permissions, and then a data
migration to rename the original permissions (if they exist) to the
new permissions. This appears to work ok. We can't easily test this
because the django-test-migrations package mutes signals during the
migrations, so no ContentTypes or Permissions end up being created.