Add a unit test that fails due to this bug
The bug occurs when a workspace data object has a second foreign
key to workspace. The inlineformset_factory method does not know
which fk it should be referring to. Add a custom workspace data
model with this foreign key. Add an adapter and a form. Add a test
for WorkspaceCreate where the exception is raised.