Remove Workspace has_access method
This method doesn't really make sense, since it is user-based
instead of group-based. A user can still have access to a workspace
if none of the groups they are in are both in the auth domain and
have the worksapce shared with them -- for example, if they are in
one group that is in the auth domain and in a second group that
the workspace is shared with.