Create both a view and staff view user in Workspace List tests
This is easier to use in tests - we don't have to create the view
user each time we need it. For tests checking information visible
to both the view user and the staff_view user, use the view_user
in the test. Only use the staff_view user when we are checking
something specific to staff_view users. (This last bit is only
relevant for the WorkspaceList view, since WorkspaceListByType is
not open to all yet.)