Add management command to migrate UUID fields
The combination of Django 5.0+ and MariaDB 10.7+ changes how UUID
fields are stored and requires handling the fields differently.
Because this app does not make an assumption about the database
backend, it doesn't make sense to redefine the fields in models.py
as suggested by Django. Instead, add a management command that can
be run by users after migrating to both Django 5.0 and MariaDb 10.7+.
The code for this command is pulled almost entirely directly from
Wagtail. See PR and related issue:
https://github.com/wagtail/wagtail/pull/11912
https://github.com/wagtail/wagtail/issues/11776