Save managed group before calling after_anvil_create
This fixes a bug where the managed group was not saved when
after_anvil_create was called, which prevented the adapter method
from using the managed group in a foreign key relationship for
another model (e.g., creating a membership using that group). To fix
this, I had to convert the CreateView to a FormView to prevent the
history from having two records (a save plus an update/save). Add
a test to catch the bug.