Remove tests for None for account adapter default values
Some of the new account adapter default values are set in the base
adapter. Therefore, it doesn't make sense to set the attribute to
None because a user subclassing the base class would like just want
the default values. In the future it might be a better idea to
change the way the account adapter behaves, and use accessor methods
to get all of the attributes instead of the attribute itself. In
addition, we might want to encourage the user to subclass the
DefaultAdapter instead of the Base adapter classes so that these
defaults get set; they still have the option of using the Base
class if they want to set everything. That can be done in a different
branch.