Change the WorkspaceClone view to use the adapter Workspace form class
Previously, the app defined a separate form to use in the WorkspaceClone
view. If a user specified a custom workspace form in the adapter,
the WorkspaceClone view would ignore that custom form when creating
a new workspace of that type, so any custom cleaning was lost. Fix
this by changing the WorkspaceCloneForm to a mixin and defining a
new form class on the fly in the WorkspaceClone view. This new class
subclasses the workspace form specified by the adapter as well as the
WorkspaceCloneFormMixin.