Rework WorkspaceCloneFormMixin to handle inheritance better
Move the authorization domain validation to the clean method, instead
of having a separate clean_authorization_domains method. This is likely
better for inheritance, because people could add their own method for
clean_authorization_domains and would be less likely to call the
super() method. Forms typically should call the super().clean method
in their clean method, and it is shown in django docs, so that is a
better place to add the custom auth domain checks.