Allow users to link an existing (unlinked) Account
If an Account already exists in the app with the same email the
user entered, and that Account is not linked to another user, allow
the user to try to link it. Previously, the view redirected with
an error message because the account already existed in the app.
This prevents a user from linking an Account that had previously been
imported by CC staff.