Do not allow a user to link a previously-linked account
I updated the code to allow a user to link an account that already
existed in the app, but that was intended for accounts that had been
imported by the CC and not linked to another user. This time, make
another update to the code and tests to prevent a user from linking
an account that had prevoiusly been linked to another user, by
querying the "AccountUserArchive" table for relevant records.