Allow users to verify previously-unlinked existing accounts
Users can now verify an account that already exists in the app, but
has never been linked to a user before. If the account is already
deactivated, it will remain deactivated until a CC staff member
reactivates it. Users are will receive an error message when they
try to link an account that has previously been linked to another
user.