Add IgnoredResult class and handlingin base AnVILAudit class
The new IgnoredResult class holds an audit results from one of the
Ignored model tables (right now there is only one). Update the
AnVILAudit class to have an attribute list that stores ignored results,
to be able to add these results, and to export them in json format
using the export method. An audit is still considered "ok" when there
are "ignored" results.