Add link to ignore create view for ManagedGroupMembership not in app results
Define a new class that subclasses NotInAppResult specifically for
ManagedGroupMembershipAudit results. This new class also stores the
group, the email, and the role that a not_in_app audit result found.
Also add a new NotInApp table for the ManagedGroupMembership audit that
shows these records, plus a link to the view to create a new
IgnoredAuditManagedGroupMembership record for that group and email.