Add ignore reporting to audit management command
Report the number of ignored records associated with an audit. To
do this, we need to tell the audit which model to look at; we can't
just use the audit results because sometimes the ignored records are
nested below (e.g., from the ManagedGroupMembershipAudit, which is
run by ManagedGroupAudit, and ManagedGroupAudit does not know about
anything ignored by ManagedGroupMembershipAudit). Perhaps there is
a better way to structure this, but we'll leave that as a future problem
to solve.